requirements.txt
matplotlib==3.5.1
numpy==1.21.2
opencv-python==4.5.5.64
Pillow==8.4.0
pytorch_lightning==1.5.7
scikit_learn==1.0.2
scipy==1.7.3
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
